refactor cell action files
