+++ x-pack/plugins/security_solution/public/common/lib/cell_actions/default_cell_actions.test.tsx
@@ -13,8 +13,8 @@ import { TGridCellAction } from '@kbn/timelines-plugin/common/types';
 import { Ecs } from '../../../../common/ecs';
 import { ColumnHeaderType } from '../../../timelines/store/timeline/model';
 
-import { defaultCellActions, EmptyComponent } from './default_cell_actions';
-import { COLUMNS_WITH_LINKS } from './helpers';
+import { defaultCellActions } from './default_cell_actions';
+import { COLUMNS_WITH_LINKS, EmptyComponent } from './helpers';
 
 describe('default cell actions', () => {
   const browserFields: BrowserFields = {};
